Added menu and in game music
Added music to the menu and game. When the MenuState and GameState
are constructed the music starts playing. When in PauseState the
music is paused and when back to game or menu the music resumes.
Note SFML does not support mp3 format so need to use something like
Audacity to export files to .ogg.
Note that you need to add sfml-audio-d.lib to Properties->Linker->
Input->Additional Dependencies otherwise you will get a bunch of
linker errors.